requirements.txt
pandas
numpy 
boto3
certifi
gensim
tqdm
urllib3
python-dotenv
konlpy
